+++ warehouse/router_identities.go
@@ -440,7 +440,7 @@ func (r *router) populateHistoricIdentities(ctx context.Context, warehouse model
 		}
 		defer whManager.Cleanup(ctx)
 
-		err = job.schemaHandle.fetchSchemaFromWarehouse(ctx, whManager)
+		_, _, err = job.schemaHandle.FetchSchemaFromWarehouse(ctx, whManager)
 		if err != nil {
 			r.logger.Errorf(`[WH]: Failed fetching schema from warehouse: %v`, err)
 			job.setUploadError(err, model.Aborted)
